+++ src/views/Settings/Appearance.Settings.vue
@@ -1,6 +1,6 @@
 <template>
   <SettingView id="appearance-settings" :title="$t('settings.appearance.title')">
-    <div id="appearance-topbar-style-section">
+    <div id="appearance-settings-content">
       <section>
         <h4 class="text-sm font-bold uppercase text-gray-3 mb-2" v-text="$t('settings.appearance.titlebarStyle')"/>
         <RadioForm